+++ lib/src/command.dart
@@ -1,4 +1,4 @@
-// action.dart
+// command.dart
 
 import 'callable.dart';
 import 'creator.dart';
@@ -11,8 +11,8 @@ import 'store.dart';
 /// Or in other words, the callback carries the [Event] to the [Store].
 /// Can be assigned to Widget properties of type [VoidCallback].
 /// The type parameter `S` is the type of the state of the [Store].
-class Action<S> extends Callable<void> {
-  const Action(this.processor, this.event);
+class Command<S> extends Callable<void> {
+  const Command(this.processor, this.event);
 
   /// The store to whose method [process](Store.process)
   /// the [event] is passed when the method [call] is called.
@@ -36,7 +36,7 @@ class Action<S> extends Callable<void> {
   /// [store] and [event] should have value semantics.
   @override
   operator ==(other) =>
-      other is Action && event == other.event && processor == other.processor;
+      other is Command && event == other.event && processor == other.processor;
 
   @override
   toString() => '$event@$processor}';
@@ -48,8 +48,8 @@ class Action<S> extends Callable<void> {
 ///
 /// The type parameter `S` is the type of the state of the [Store].
 /// The type parameter `V` is the type of the value of the [Event1].
-class Action1<S, V> extends Callable1<void, V> {
-  const Action1(this.processor, this.event);
+class Command1<S, V> extends Callable1<void, V> {
+  const Command1(this.processor, this.event);
 
   /// The store to whose method [process](Store.process)
   /// the [event] is passed when the method [call] is called.
@@ -73,7 +73,7 @@ class Action1<S, V> extends Callable1<void, V> {
   /// [store] and [event] should have value semantics.
   @override
   operator ==(other) =>
-      other is Action1 && event == other.event && processor == other.processor;
+      other is Command1 && event == other.event && processor == other.processor;
 
   @override
   toString() => '$event@$processor}';
@@ -86,8 +86,8 @@ class Action1<S, V> extends Callable1<void, V> {
 /// The type parameter `S` is the type of the state of the [Store].
 /// The type parameter `V1` is the type of the 1st value of the [Event2].
 /// The type parameter `V2` is the type of the 2nd value of the [Event2].
-class Action2<S, V1, V2> extends Callable2<void, V1, V2> {
-  const Action2(this.processor, this.event);
+class Command2<S, V1, V2> extends Callable2<void, V1, V2> {
+  const Command2(this.processor, this.event);
 
   /// The store to whose method [process](Store.process)
   /// the [event] is passed when the method [call] is called.
@@ -112,7 +112,7 @@ class Action2<S, V1, V2> extends Callable2<void, V1, V2> {
   /// [store] and [event] should have value semantics.
   @override
   operator ==(other) =>
-      other is Action2 && event == other.event && processor == other.processor;
+      other is Command2 && event == other.event && processor == other.processor;
 
   @override
   toString() => '$event@$processor}';
@@ -126,8 +126,8 @@ class Action2<S, V1, V2> extends Callable2<void, V1, V2> {
 /// The type parameter `V1` is the type of the 1st value of the [Event3].
 /// The type parameter `V2` is the type of the 2nd value of the [Event3].
 /// The type parameter `V3` is the type of the 3rd value of the [Event3].
-class Action3<S, V1, V2, V3> extends Callable3<void, V1, V2, V3> {
-  const Action3(this.processor, this.event);
+class Command3<S, V1, V2, V3> extends Callable3<void, V1, V2, V3> {
+  const Command3(this.processor, this.event);
 
   /// The store to whose method [process](Store.process)
   /// the [event] is passed when the method [call] is called.
@@ -152,7 +152,7 @@ class Action3<S, V1, V2, V3> extends Callable3<void, V1, V2, V3> {
   /// [store] and [event] should have value semantics.
   @override
   operator ==(other) =>
-      other is Action3 && event == other.event && processor == other.processor;
+      other is Command3 && event == other.event && processor == other.processor;
 
   @override
   toString() => '$event@$processor}';
@@ -164,14 +164,14 @@ class Action3<S, V1, V2, V3> extends Callable3<void, V1, V2, V3> {
 ///
 /// The type parameter `S` is the type of the state of the [Store].
 /// The type parameter `R` is the type of the future.
-class FutureAction<S, R> extends Callable<void> {
+class FutureCommand<S, R> extends Callable<void> {
   final EventProcessor<S> processor;
   final Event<S>? onStarted;
   final FutureCreator<R> creator;
   final Event1<S, R> onValue;
   final ErrorEvent<S>? onError;
 
-  FutureAction({
+  FutureCommand({
     required this.processor,
     required this.creator,
     this.onStarted,
@@ -210,7 +210,7 @@ class FutureAction<S, R> extends Callable<void> {
 
   @override
   operator ==(other) =>
-      other is FutureAction<S, R> &&
+      other is FutureCommand<S, R> &&
       processor == other.processor &&
       creator == other.creator &&
       onStarted == other.onStarted &&
@@ -225,14 +225,14 @@ class FutureAction<S, R> extends Callable<void> {
 /// The type parameter `S` is the type of the state of the [Store].
 /// The type parameter `R` is the type of the future.
 /// The type parameter `P` is the type of the parameter that is transfered form the callable to the  future creator.
-class FutureAction1<S, R, P> extends Callable1<void, P> {
+class FutureCommand1<S, R, P> extends Callable1<void, P> {
   final EventProcessor<S> processor;
   final FutureCreator1<R, P> creator;
   final Event<S>? onStarted;
   final Event1<S, R> onValue;
   final ErrorEvent<S>? onError;
 
-  FutureAction1({
+  FutureCommand1({
     required this.processor,
     this.onStarted,
     required this.creator,
@@ -268,7 +268,7 @@ class FutureAction1<S, R, P> extends Callable1<void, P> {
 
   @override
   operator ==(other) =>
-      other is FutureAction1<S, R, P> &&
+      other is FutureCommand1<S, R, P> &&
       processor == other.processor &&
       creator == other.creator &&
       onStarted == other.onStarted &&
@@ -284,14 +284,14 @@ class FutureAction1<S, R, P> extends Callable1<void, P> {
 /// The type parameter `R` is the type of the future.
 /// The type parameter `P1` is the type of the 1st parameter that is transfered form the callable to the  future creator.
 /// The type parameter `P2` is the type of the 2nd parameter that is transfered form the callable to the  future creator.
-class FutureAction2<S, R, P1, P2> extends Callable2<void, P1, P2> {
+class FutureCommand2<S, R, P1, P2> extends Callable2<void, P1, P2> {
   final EventProcessor<S> processor;
   final FutureCreator2<R, P1, P2> creator;
   final Event<S>? onStarted;
   final Event1<S, R> onValue;
   final ErrorEvent<S>? onError;
 
-  FutureAction2({
+  FutureCommand2({
     required this.processor,
     this.onStarted,
     required this.creator,
@@ -327,7 +327,7 @@ class FutureAction2<S, R, P1, P2> extends Callable2<void, P1, P2> {
 
   @override
   operator ==(other) =>
-      other is FutureAction2<S, R, P1, P2> &&
+      other is FutureCommand2<S, R, P1, P2> &&
       processor == other.processor &&
       creator == other.creator &&
       onStarted == other.onStarted &&
@@ -341,7 +341,7 @@ class FutureAction2<S, R, P1, P2> extends Callable2<void, P1, P2> {
 ///
 /// The type parameter `S` is the type of the state of the [Store].
 /// The type parameter `R` is the type of the stream.
-class StreamAction<S, R> extends Callable<void> {
+class StreamCommand<S, R> extends Callable<void> {
   final EventProcessor<S> processor;
   final StreamCreator<R> creator;
   final Event<S>? onStarted;
@@ -349,7 +349,7 @@ class StreamAction<S, R> extends Callable<void> {
   final Event<S>? onDone;
   final ErrorEvent<S>? onError;
 
-  StreamAction({
+  StreamCommand({
     required this.processor,
     this.onStarted,
     required this.creator,
@@ -386,7 +386,7 @@ class StreamAction<S, R> extends Callable<void> {
 
   @override
   operator ==(other) =>
-      other is StreamAction<S, R> &&
+      other is StreamCommand<S, R> &&
       processor == other.processor &&
       creator == other.creator &&
       onStarted == other.onStarted &&
@@ -402,7 +402,7 @@ class StreamAction<S, R> extends Callable<void> {
 /// The type parameter `S` is the type of the state of the [Store].
 /// The type parameter `R` is the type of the future.
 /// The type parameter `P` is the type of the parameter that is transfered form the callable to the stream creator.
-class StreamAction1<S, R, P> extends Callable1<void, P> {
+class StreamCommand1<S, R, P> extends Callable1<void, P> {
   final EventProcessor<S> processor;
   final StreamCreator1<R, P> creator;
   final Event<S>? onStarted;
@@ -410,7 +410,7 @@ class StreamAction1<S, R, P> extends Callable1<void, P> {
   final Event<S>? onDone;
   final ErrorEvent<S>? onError;
 
-  StreamAction1({
+  StreamCommand1({
     required this.processor,
     this.onStarted,
     required this.creator,
@@ -447,7 +447,7 @@ class StreamAction1<S, R, P> extends Callable1<void, P> {
 
   @override
   operator ==(other) =>
-      other is StreamAction1<S, R, P> &&
+      other is StreamCommand1<S, R, P> &&
       processor == other.processor &&
       creator == other.creator &&
       onStarted == other.onStarted &&
@@ -464,7 +464,7 @@ class StreamAction1<S, R, P> extends Callable1<void, P> {
 /// The type parameter `R` is the type of the future.
 /// The type parameter `P1` is the type of the 1st parameter that is transfered form the callable to the stream creator.
 /// The type parameter `P2` is the type of the 2nd parameter that is transfered form the callable to the stream creator.
-class StreamAction2<S, R, P1, P2> extends Callable2<void, P1, P2> {
+class StreamCommand2<S, R, P1, P2> extends Callable2<void, P1, P2> {
   final EventProcessor<S> processor;
   final StreamCreator2<R, P1, P2> creator;
   final Event<S>? onStarted;
@@ -472,7 +472,7 @@ class StreamAction2<S, R, P1, P2> extends Callable2<void, P1, P2> {
   final Event<S>? onDone;
   final ErrorEvent<S>? onError;
 
-  StreamAction2({
+  StreamCommand2({
     required this.processor,
     this.onStarted,
     required this.creator,
@@ -509,7 +509,7 @@ class StreamAction2<S, R, P1, P2> extends Callable2<void, P1, P2> {
 
   @override
   operator ==(other) =>
-      other is StreamAction2<S, R, P1, P2> &&
+      other is StreamCommand2<S, R, P1, P2> &&
       processor == other.processor &&
       creator == other.creator &&
       onStarted == other.onStarted &&