+++ lib/src/function.dart
@@ -8,21 +8,23 @@ import 'store.dart';
 
 /// A function that maps store date to props.
 ///
-/// The type parameter `S` is the type of the [StoreData].
+/// The type parameter `S` is the type of the [StoreSnapshot].
 /// The type parameter `P` is the return type of the function.
-typedef StoreDataToPropsMapper<S, P> = P Function(StoreData<S> storeData,
-    [String? routeName]);
+typedef SnapshotToPropsMapper<S, P> = P Function(
+  StoreSnapshot<S> snapshot,
+  String? routeName,
+);
 
 /// A function that maps double store data to props.
 ///
-/// The type parameter `S1` is the type of the 1st [StoreData].
-/// The type parameter `S2` is the type of the 2nd [StoreData].
+/// The type parameter `S1` is the type of the 1st [StoreSnapshot].
+/// The type parameter `S2` is the type of the 2nd [StoreSnapshot].
 /// The type parameter `P` is the return type of the function.
-typedef DoubleStoreDataToPropsMapper<S1, S2, P> = P Function(
-  StoreData<S1> appStoreData,
-  StoreData<S2> pageStoreData, [
+typedef SnapshotsToPropsMapper<S1, S2, P> = P Function(
+  StoreSnapshot<S1> snapshot1,
+  StoreSnapshot<S2> snapshot2,
   String? routeName,
-]);
+);
 
 // A function that builds a Widget from a props parameter.
 ///
@@ -35,13 +37,14 @@ typedef WidgetFromPropsBuilder<P> = Widget Function({
 /// A function that can be registered as an event listener at a Store.
 ///
 /// A registered listener is called by the Store at the end of each dispatch method execution.
+/// The type parameter `S` is the type of the Event.
 typedef EventListener<S> = void Function(
-  StoreData<S> processor,
+  StoreSnapshot<S> snapshot,
   Event<S> event,
   UniqueKey key,
 );
 
 /// A function that accepts an Event as parameter.
 ///
-/// The type parameter `P` is the type of the Event.
+/// The type parameter `S` is the type of the Event.
 typedef EventAcceptor<S> = void Function(Event<S> event);