+++ test/command/command1_test.dart
@@ -1,13 +1,13 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:reduced/src/action.dart';
+import 'package:reduced/src/command.dart';
 import 'package:reduced/src/event.dart';
 import 'package:reduced/src/store.dart';
 
-class MockEvent3 extends Event3<String, String, String, String> {
-  MockEvent3();
+class MockEvent1 extends Event1<String, String> {
+  MockEvent1();
 
   @override
-  call(state, value1, value2, value3) => '$state $value1 $value2 $value3';
+  call(state, value) => '$state $value';
 }
 
 class MockStore extends Store<String> {
@@ -17,38 +17,38 @@ class MockStore extends Store<String> {
   String state;
 
   @override
-  process(event) => state = event(state);
+  process(reducer) => state = reducer(state);
 }
 
 void main() {
-  test('Action3 call test', () {
+  test('Action1 call test', () {
     final store = MockStore('0');
-    final objectUnterTest = Action3(
+    final objectUnterTest = Command1(
       store,
-      MockEvent3(),
+      MockEvent1(),
     );
-    objectUnterTest.call('1', '2', '3');
-    expect(store.state, '0 1 2 3');
+    objectUnterTest.call('1');
+    expect(store.state, '0 1');
   });
 
-  test('Action3 hashCode test', () {
+  test('Action1 hashCode test', () {
     final store1 = MockStore('1');
-    final reducer1 = MockEvent3();
+    final reducer1 = MockEvent1();
     final store2 = MockStore('2');
-    final reducer2 = MockEvent3();
-    final objectUnterTest11 = Action3(
+    final reducer2 = MockEvent1();
+    final objectUnterTest11 = Command1(
       store1,
       reducer1,
     );
-    final objectUnterTest12 = Action3(
+    final objectUnterTest12 = Command1(
       store1,
       reducer2,
     );
-    final objectUnterTest21 = Action3(
+    final objectUnterTest21 = Command1(
       store2,
       reducer1,
     );
-    final objectUnterTest22 = Action3(
+    final objectUnterTest22 = Command1(
       store1,
       reducer1,
     );
@@ -57,24 +57,24 @@ void main() {
     expect(objectUnterTest11.hashCode, isNot(objectUnterTest21.hashCode));
     expect(objectUnterTest12.hashCode, isNot(objectUnterTest21.hashCode));
   });
-  test('Action3 operator== test', () {
+  test('Action1 operator== test', () {
     final store1 = MockStore('1');
-    final reducer1 = MockEvent3();
+    final reducer1 = MockEvent1();
     final store2 = MockStore('2');
-    final reducer2 = MockEvent3();
-    final objectUnterTest11 = Action3(
+    final reducer2 = MockEvent1();
+    final objectUnterTest11 = Command1(
       store1,
       reducer1,
     );
-    final objectUnterTest12 = Action3(
+    final objectUnterTest12 = Command1(
       store1,
       reducer2,
     );
-    final objectUnterTest21 = Action3(
+    final objectUnterTest21 = Command1(
       store2,
       reducer1,
     );
-    final objectUnterTest22 = Action3(
+    final objectUnterTest22 = Command1(
       store1,
       reducer1,
     );